Reduce some code duplication in the Joining namespace
All child classes have score_mode so we can move it to the abstract base class
